Fix TIMERTIME_T_MAX sign warning.

+++ src/system/timer.h
@@ -46,7 +46,7 @@ typedef struct TimerEvent_s
  */
 #ifndef TimerTime_t
 typedef uint32_t TimerTime_t;
-#define TIMERTIME_T_MAX (~0)
+#define TIMERTIME_T_MAX                             ( ( uint32_t )~0 )
 #endif
 
 /*!
